xfrm: Introduce 'dir' attribute for SA
Inspired by the upcoming IP-TFS patch set, and confusions experienced in
the past due to lack of direction on SAs, this new direction attribute
aims to streamline the SA configuration process and enhance the clarity
of existing SA attributes.

This patch set introduces the 'dir' attribute to SA, aka xfrm_state,
('in' for input or 'out' for output). It adds validation of existing
direction-specific SA attributes during configuration and in the data
path lookup.

This change should not affect any existing use case or way of configuring
SA. You will notice improvements when the new 'dir' attribute is set.
--
    v10->v11:
     - set dir out for acquire state
     - allow output state with ESN to set replay-window 0

    v9->v10:
     - add more direction specific validations
            XFRM_STATE_NOPMTUDISC, XFRM_SA_XFLAG_DONT_ENCAP_DSCP
            XFRMA_MTIMER_THRESH
     - refactor validations into a fuction.
     - add dir to ALLOCSPI to support strongSwan updating SPI to "in" state

    v8->v9:
     - add validation XFRM_STATE_ICMP not allowed on OUT SA.

    v7->v8:
     - add extra validation check on replay window and seq
     - XFRM_MSG_UPDSA old and new SA should match "dir"

    v6->v7:
     - add replay-window check non-esn 0 and ESN 1.
     - remove :XFRMA_SA_DIR only allowed with HW OFFLOAD

    v5->v6:
     - XFRMA_SA_DIR only allowed with HW OFFLOAD

    v4->v5:
     - add details to commit message

    v3->v4:
     - improve HW OFFLOAD DIR check add the other direction

    v2->v3:
     - delete redundant XFRM_SA_DIR_USE
     - use u8 for "dir"
     - fix HW OFFLOAD DIR check

    v1->v2:
     - use .strict_start_type in struct nla_policy xfrma_policy
     - delete redundant XFRM_SA_DIR_MAX enum

Signed-off-by: Antony Antony <antony.antony@secunet.com>