xfrm: Introduce direction attribute for SA
Hi,

Inspired by the upcoming IP-TFS patch set, and confusions experienced in
the past due to lack of direction attribute on SAs, add a new direction
"dir" attribute. It aims to streamline the SA configuration process and
enhance the clarity of existing SA attributes.

This patch set introduces the 'dir' attribute to SA, aka xfrm_state,
('in' for input or 'out' for output). Alsp add validations of existing
direction-specific SA attributes during configuration and in the data
path lookup.

This change would not affect any existing use case or way of configuring
SA. You will notice improvements when the new 'dir' attribute is set.

Antony Antony (4):
  xfrm: Add Direction to the SA in or out
  xfrm: Add dir validation to "out" data path lookup
  xfrm: Add dir validation to "in" data path lookup
  xfrm: Restrict SA direction attribute to specific netlink message types

 include/net/xfrm.h        |   1 +
 include/uapi/linux/snmp.h |   2 +
 include/uapi/linux/xfrm.h |   6 ++
 net/ipv6/xfrm6_input.c    |   7 ++
 net/xfrm/xfrm_compat.c    |   7 +-
 net/xfrm/xfrm_device.c    |   6 ++
 net/xfrm/xfrm_input.c     |  11 +++
 net/xfrm/xfrm_policy.c    |   6 ++
 net/xfrm/xfrm_proc.c      |   2 +
 net/xfrm/xfrm_replay.c    |   3 +-
 net/xfrm/xfrm_state.c     |   5 ++
 net/xfrm/xfrm_user.c      | 147 ++++++++++++++++++++++++++++++++++++--
 12 files changed, 194 insertions(+), 9 deletions(-)

--
2.30.2