net: add dev_noqueue_xmit_list() helper function
***This replaces validate_xmit_skb() by validate_xmit_skb_list()***

This new helper function wraps the device has no queue case.  This function can
be reused from listified skb in the tx path since the device has no queue path
already supports for skb list.

An alternative to this patch to reuse a smaller fraction of code can be to wrap
this common code instead in a function:

+                       HARD_TX_LOCK(dev, txq, cpu);
+
+                       if (!netif_xmit_stopped(txq)) {
+                               dev_xmit_recursion_inc();
+                               skb = dev_hard_start_xmit(skb, dev, txq, &rc);
+                               dev_xmit_recursion_dec();
+                               if (dev_xmit_complete(rc)) {
+                                       HARD_TX_UNLOCK(dev, txq);
+                                       goto out;
+                               }
+                       }
+                       HARD_TX_UNLOCK(dev, txq);

Note: This comment was in the original patch from Steffen:

       /* FIXME: For each skb!!! */
       dev_core_stats_tx_dropped_inc(dev);
       kfree_skb_list(skb);

but probably upstream needs a fix for this too given that this path already
supports for skb list. Then, probably not worth to worry about this in this
series.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>